RA-20 refactoring
