+++ app/src/main/java/com/renegatemaster/recipeapp/ui/recipes/favorites/FavoritesFragment.kt
@@ -1,4 +1,4 @@
-package com.renegatemaster.recipeapp.ui
+package com.renegatemaster.recipeapp.ui.recipes.favorites
 
 import android.content.Context
 import android.os.Bundle
@@ -13,6 +13,8 @@ import androidx.fragment.app.replace
 import com.renegatemaster.recipeapp.R
 import com.renegatemaster.recipeapp.data.STUB
 import com.renegatemaster.recipeapp.databinding.FragmentFavoritesBinding
+import com.renegatemaster.recipeapp.ui.recipes.recipe.RecipeFragment
+import com.renegatemaster.recipeapp.ui.recipes.recipe_list.RecipesListAdapter
 import com.renegatemaster.recipeapp.utils.Constants
 
 class FavoritesFragment: Fragment() {
