+++ app/src/main/res/layout/fragment_favorites.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/backgroundMain"
     android:orientation="vertical"
-    tools:context=".ui.FavoritesFragment">
+    tools:context=".ui.recipes.favorites.FavoritesFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
