+++ app/src/main/res/layout/fragment_list_recipes.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/backgroundMain"
     android:orientation="vertical"
-    tools:context=".ui.RecipesListFragment">
+    tools:context=".ui.recipes.recipe_list.RecipesListFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
