+++ app/src/main/res/layout/fragment_recipe.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/backgroundMain"
-    tools:context=".ui.RecipeFragment">
+    tools:context=".ui.recipes.recipe.RecipeFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
